refactored responses code, cleaned up folder structure
